+++ CHANGES.md
@@ -20,3 +20,4 @@
 -   Styled `a` text to be white;
 -   Tweaked footer links left/right spacing
 -   Installed `@gov.au/body`
+-   Added external link icon to footer